move NTP set time to Settings page
